SERVER-5140: avoid a bug in SCons with subprocess.Popen
SCons overrides the subprocess module with its own custom implementation
(forked from an earlier version of Python, but with some patches).
SCons' implementation acquires a lock but does not release it during
Popen() if the command being executed cannot be found. The next
invocation of Popen() will hang indefinitely attempting to acquire the
lock.

Moving smoke_python_name() to a python module (i.e. out of the
SCons-managed files) causes "import subprocess" to find the actual,
system-installed subprocess module, which does not contain the buggy
locking.